api/requirements.txt
fastjsonschema==2.16.2
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Limiter==2.7.0
Flask-SQLAlchemy==2.5.1
jedi==0.18.1
jh-utils==0.1.40
numpy==1.23.3
pandas==1.2
pip==22.2.2
psycopg2==2.8.6
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
setuptools==65.5.0
six==1.16.0
SQLAlchemy==1.4.39
stack-data==0.6.0
traitlets==5.5.0
urllib3==1.26.12
websocket-client==1.4.2
Werkzeug==2.0.3
pyvirtualdisplay==2.2
Pillow==8.3.2